Move web registry integration test to web/
